SchedulerItem skipEvent support added

+++ src/SchedulerItem.cpp
@@ -3,7 +3,7 @@
 SchedulerItem::SchedulerItem() 
     : BaseItem(), start(Time(0, 0)), duration(Time(0, 0)), weekday("0000000"), enabled(false), mode(""), skipped(false) {}
 
-SchedulerItem::SchedulerItem(uint64_t id, uint64_t event_id, std::string name, Time start, Time duration, std::string weekday, bool enabled, bool on, std::string mode, bool skipped)
+SchedulerItem::SchedulerItem(uint64_t id, short event_id, std::string name, Time start, Time duration, std::string weekday, bool enabled, bool on, std::string mode, bool skipped)
     :BaseItem(id, event_id, name, on), start(start), duration(duration), weekday(weekday), enabled(enabled), mode(mode), skipped(skipped) {}
 
 Time SchedulerItem::getStartTime() 
@@ -60,7 +60,7 @@ bool SchedulerItem::isSkipped()
     return skipped;
 }
 
-uint64_t SchedulerItem::getSkipEventId()
+short SchedulerItem::getSkipEventId()
 {
     return skipEvent_id;
 }
@@ -73,7 +73,7 @@ void SchedulerItem::populateDerivedClassFromJson(JsonDocument &doc)
     this->enabled = doc["enabled"].as<bool>();
     this->mode = doc["mode"].as<std::string>();
     this->skipped = doc["skipped"].as<bool>();
-    this->skipEvent_id = doc["skipEvent_id"].as<uint64_t>();
+    this->skipEvent_id = doc["skipEvent_id"].as<short>();
 }
 
 void SchedulerItem::derivedClassToJson(JsonDocument &doc)
